+++ test/globalize_test.rb
@@ -28,6 +28,17 @@ class GlobalizeTest < MiniTest::Unit::TestCase
     end
   end
 
+  test "should set friendly id for locale" do
+    transaction do
+      article = TranslatedArticle.create!(:title => "War and Peace")
+      article.set_friendly_id("Guerra y paz", :es)
+      article.save!
+      article = TranslatedArticle.find('war-and-peace')
+      I18n.with_locale(:es) { assert_equal "guerra-y-paz", article.friendly_id }
+      I18n.with_locale(:en) { assert_equal "war-and-peace", article.friendly_id }
+    end
+  end
+
   # https://github.com/svenfuchs/globalize3/blob/master/test/globalize3/dynamic_finders_test.rb#L101
   # see: https://github.com/svenfuchs/globalize3/issues/100
   test "record returned by friendly_id should have all translations" do